requirements.txt
torch
torchvision
streamlit
lightning
# openai_clip
open_clip_torch